refactor(shared): consolidate telemetry helpers into telemetryFormat
Following mcharles' point about telemetry formatters living in
`utility.ts` alongside unrelated helpers (deepClone, debounce,
downloadBlob), pull the entire telemetry surface into
`@/shared/utils/telemetryFormat` so there's one home for unit math
and display strings.

Moved into `telemetryFormat`:
- Unit conversions: convertMegahashSecToTerahashSec,
  convertGigahashSecToTerahashSec, convertWtoKW, convertCtoF,
  convertFtoC
- Hashrate ladder: formatHashrateWithUnit + the TH_TO_PH_THRESHOLD /
  TH_TO_PH_DIVISOR constants the dashboard chart code uses
- Temperature: formatTempRange, getAsicTempValue

Already in `telemetryFormat` (from the prior refactor):
- formatHashrate / formatHashrateOrDash
- formatPowerUsedCapacity / formatPowerMwOrDash
- formatEfficiency / formatEfficiencyOrDash
- KW_PER_MW

Added formatPowerKwOrDash so rack-level surfaces have a kW-scoped
formatter (sites/buildings use MW) and don't have to inline
`${getDisplayValue(...)} kW` themselves.

`utility.ts` is now left with the non-telemetry helpers only.

Callers updated across protoFleet, protoOS, and shared (~12 files):
HashRateValue, AsicPopover, TemperatureValue, MinerTemperature,
HashratePanel, DeviceSetPerformanceSection, deviceSetColConfig,
PowerTargetPopover, telemetryUtils, kpis/hooks utility,
diagnostic AsicPopover utility, MinerList — all point at the new
telemetryFormat module.

`rackCardMapper.formatRackCardStats` now uses the shared formatters
(formatHashrate, formatEfficiency, formatPowerKwOrDash) instead of
inline `${getDisplayValue(...)} TH/s` — picks up the auto-scaling
unit ladder (so "2,500,000 TH/s" reads as "2.50 PH/s" on rack cards)
and aligns with the other rollup surfaces.

Tests: utility.test.ts loses the formatHashrateWithUnit /
formatTempRange suites; same tests moved to telemetryFormat.test.ts.
All 2522 client tests pass.
