+++ client/src/shared/utils/utility.ts
@@ -1,5 +1,4 @@
-import type { TemperatureUnit } from "@/shared/features/preferences";
-import { getDisplayValue, padLeft } from "@/shared/utils/stringUtils";
+import { padLeft } from "@/shared/utils/stringUtils";
 
 export const deepClone = (obj: any) => {
   const stringify = JSON.stringify(obj, (_, value) => (typeof value === "bigint" ? Number(value) : value));
@@ -44,62 +43,9 @@ export const getRandomFloat = (min: number, max: number, precision: number = 100
   );
 };
 
-export const convertMegahashSecToTerahashSec = (value?: number | null) => (value ?? 0) / 1000000;
-export const convertGigahashSecToTerahashSec = (value?: number | null) => (value ?? 0) / 1000;
-export const convertWtoKW = (value?: number | null) => (value ?? 0) / 1000;
-
-// Hashrate unit conversion constants. Retained for backwards compatibility
-// with callers that import these directly; formatHashrateWithUnit no longer
-// uses TH_TO_PH_THRESHOLD/TH_TO_PH_DIVISOR internally.
-export const TH_TO_PH_THRESHOLD = 1000;
-export const TH_TO_PH_DIVISOR = 1000;
-
-// Unit ladder for hashrate display. Value in is always TH/s; the formatter
-// scales it into GH/s, TH/s, PH/s, or EH/s so the displayed number stays in
-// [1, 1000) for non-zero inputs. Zero stays in TH/s as the conventional
-// "no signal" default.
-const TH_PER_PH = 1_000;
-const TH_PER_EH = 1_000_000;
-const GH_PER_TH = 1_000;
-
-export const formatHashrateWithUnit = (value: number = 0) => {
-  // NaN/Infinity guard. Bad upstream data (corrupted telemetry, division
-  // by zero in an aggregation) shouldn't propagate `NaN TH/s` to the UI;
-  // fall through to the zero rendering instead.
-  if (!Number.isFinite(value)) {
-    return { value: 0, unit: "TH/S" };
-  }
-  if (value <= 0) {
-    return { value: 0, unit: "TH/S" };
-  }
-  // Strict `>` boundaries keep prior callers (AsicPopover, HashRateValue)
-  // rendering 1000 TH/s as "1000 TH/S" instead of "1 PH/S"; same rule
-  // applies one step up at the EH boundary.
-  if (value > TH_PER_EH) {
-    return { value: value / TH_PER_EH, unit: "EH/S" };
-  }
-  if (value > TH_PER_PH) {
-    return { value: value / TH_PER_PH, unit: "PH/S" };
-  }
-  if (value < 1) {
-    return { value: value * GH_PER_TH, unit: "GH/S" };
-  }
-  return { value, unit: "TH/S" };
-};
-
-export const convertCtoF = (value: number = 0) => (value * 9) / 5 + 32;
-export const convertFtoC = (value: number = 0) => ((value - 32) * 5) / 9;
-
-export const formatTempRange = (minC: number, maxC: number, temperatureUnit: TemperatureUnit): string => {
-  const min = temperatureUnit === "F" ? convertCtoF(minC) : minC;
-  const max = temperatureUnit === "F" ? convertCtoF(maxC) : maxC;
-  return `${getDisplayValue(min)} °${temperatureUnit} – ${getDisplayValue(max)} °${temperatureUnit}`;
-};
-
-export const getAsicTempValue = (avgAsicTemp: number | undefined, isFahrenheit: boolean) => {
-  if (!avgAsicTemp) return "N/A"; // TODO: why not return undefined, so we can show skeleton, also 0 cound be falsey
-  return isFahrenheit ? convertCtoF(avgAsicTemp) : avgAsicTemp;
-};
+// Telemetry conversions + formatters (formatHashrateWithUnit, convertCtoF,
+// formatTempRange, etc.) live in @/shared/utils/telemetryFormat. Import
+// from there for telemetry display; this file keeps non-telemetry helpers.
 
 const alphabet = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
